+++ test/fizz_buzz_test.rb
@@ -1,0 +1,13 @@
+require 'minitest/autorun'
+#lib/fizz_buzz.rbを読み込む
+require_relative '../lib/fizz_buzz'
+
+class FizzBuzzTest < Minitest::Test
+  def test_fizz_buzz
+    assert_equal '1', fizz_buzz(1)
+    assert_equal '2', fizz_buzz(2)
+    assert_equal 'Fizz', fizz_buzz(3)
+    assert_equal 'Buzz', fizz_buzz(5)
+    assert_equal 'FizzBuzz', fizz_buzz(15)
+  end
+end
